refactor: CircleCheckBox 디렉터리 변경
